chore (backend): refactor code into cleaner modules and use a versioned db migration strategy (#1577)
* tmp checkpoint for db migrations

* feat (backend): support versioned migrations for sqlite db

* chore (backend): refactor js apis into separate module

* feat (backend): add basic kv store

* chore (backend): refactor messages to api module

* chore (backend): move handlers to src/api

* chore (backend): refactor ai modules under consolidated module

* chore(backend): rename backend module to worker

* chore (backend): move back handlers to worker and expose key value store js apis

* chore (backend): remove kv code for this refactor

* chore (backend): move drop tables before alter tables and skip backup db

Statements after alter table statements for add columns in the ai
sessions and ai session messages error out because of the
duplicate column error. We ignore this error but statements after that
won't be executed.

We also are temporarily skipping db backup step for this migration as
it's just the first migration file and we don't know yet how much this
might delay startup depending on how big user's databases are.